requirements.txt
python-binance==1.0.19
anthropic==0.39.0
python-telegram-bot==20.8
ta==0.11.0
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0
requests==2.31.0